api: remove residual compatibility helpers
