+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- Removed residual compatibility-only public APIs before 0.4.0. Migrate `controller.try_begin_request(route)` and `controller.try_begin_request_with(route, options)` to `controller.begin_request(route)` and `controller.begin_request_with(route, options)`; callers that truly need an admission branch can match `ControllerRequestHandle::{Active, Inert}`. Migrate `RuntimeSampler::start(run, interval)?` to `RuntimeSampler::builder(run).interval(interval).start()?`. The `crate_name()` smoke helpers in `tailtriage-tokio` and `tailtriage-axum` were removed without replacement because they were not product behavior. The `tailtriage-cli` command-line package and library documentation target remain supported, but its artifact loader, analyzer-option builder/error, and analyzer-options help renderer are now internal CLI plumbing: use `tailtriage-analyzer` for in-process analysis, `tailtriage-core` for generic Run validation/normalization, and invoke the CLI for saved-artifact command policy.
+
 - Simplified the analyzer API to one checked `analyze_run` operation plus separate `render_text`, `render_json`, and `render_json_pretty` functions. Migrate `try_analyze_run(...)` to `analyze_run(...)?`, `Analyzer` methods to the free function, `analyze_run_json*` / `try_analyze_run_json*` to analysis followed by `render_json*`, and analyzer strict wrappers to `tailtriage_core::validate_run_strict(...)?` followed by analysis.
 
 - Executor-pressure scoring now normalizes runnable queue depth by Tokio worker count when complete worker evidence is available. Historical artifacts retain exact legacy scoring, while partial, inconsistent, or invalid worker evidence uses a confidence-capped legacy fallback.
